git push googlecode master:master
Updated a few things
Removed a bug in the navigation (telling through logcat that the action is unknown when that is not the case)
Added DisplayPanel.java: Show a bitmap and text on the screen, going to be used by the media menu.

This is just a debug release, this version will not be avalable as apk.

Figured something out? Fixed a bug? Added something great? E-Mail me!
info@rnplus.nl